♻️  router runtime_preflight.py 추출 — runtime intent + join driver (P0-P step 11)
변경 이유
- ``_run_runtime_preflight`` + ``_handle_join_or_append`` + runtime
  input marshalling + clarification 디스플레이 — 모두 runtime
  classifier 가 ``CONTINUE_EXISTING_WORK`` 같은 intent 를 잡았을 때
  legacy 대신 join/append 로 흘러보내는 코드인데 ``_legacy`` 안에서
  400+ 줄. 가장 entangled 했고 P0-K/M 가드 우선순위에도 직접 영향.

주요 변경 사항
- 새 파일 ``runtime_preflight.py`` (약 530 줄):
  - ``_PREFLIGHT_SHORT_CIRCUIT_INTENTS`` 4 intent
  - ``_run_runtime_preflight`` — runtime understand/decide → 동작
  - ``_handle_join_or_append`` — legacy join helper (research_loop
    재실행 결정 포함)
  - ``_thread_id_for_runtime`` / ``_observation_for_runtime`` —
    discord message → RuntimeInput 마샬링
  - ``_format_runtime_preflight_clarification`` — 다중 후보 디스플레이
  - clarification cache / phrase_detect / runtime / obsidian_gate
    / reporting / research_loop / session_persistence / utils 모두
    explicit import.
- ``_handle_clarification_selection`` shim — 최종 구현은 step 12 에서
  main.py 로 이동. shim 으로 임시 cycle 피함.
- ``_legacy.py`` 의 해당 6 심볼 (424 줄) 제거 → re-export.
- ``__init__.py`` facade 가 ``.runtime_preflight`` + clarification
  cache 모듈 직접 import.

비고
- 4672 PASS + 1 skipped 회귀 유지.
- ``_legacy.py`` 1399 → 1010 줄. 누적 P0-P 전체 3316 → 1010 (총 2306
  줄 감소). 남은 책임은 step 12 의 main route + clarification CREATE
  driver + 일부 코딩 gate 잔여.
